Incorporated the searchbar onto the dashboard. Still need to have it redirect to the results page. Also made some style improvements.
