--- frontend/src/components/DashboardWidget.jsx
@@ -1,11 +1,0 @@
-
-function DashboardWidget({title, content, identifier}) {
-    return (
-        <div id="widget" className={identifier}>
-            <h3>{title}</h3>
-            {content}
-        </div>
-    );
-}
-
-export default DashboardWidget;
