+++ frontend/src/index.css
@@ -18,6 +18,7 @@ body {
 .App {
     display: grid;
     gap: 80px;
+    font-family: "Inria Sans", sans-serif;
     
     #navbar {
         align-self: start;
@@ -52,6 +53,7 @@ body {
 
     #nav-logo {
         align-self: center;
+        font-size: 20px;
     }
 
     #nav-logo:hover {
@@ -96,6 +98,7 @@ body {
     max-width: 1200px;
 
     .collection-financials {
+        grid-row: 1/3;
         p {
             font-size: 28px;
             font-weight: 700;
@@ -130,7 +133,7 @@ body {
             text-align: center;
             border: none;
             border-radius: 4px;
-            padding: 10px 20px;
+            padding: 6px 0;
             color: white;
             width: 60%;
             margin: 0 auto;
@@ -153,6 +156,10 @@ body {
         }
     }
 
+    .search-inventory {
+
+    }
+
     .recently-added {
         grid-column: 1/3;
 
@@ -165,7 +172,7 @@ body {
 
     .inventory-breakdown {
         grid-column: 3/4;
-        grid-row: 1/3;
+        grid-row: 1/4;
 
         #total-artifacts {
             margin-top: 20px;
@@ -190,13 +197,13 @@ body {
     background-color: rgb(222, 222, 222);
     border-radius: 12px;
     padding: 10px 20px;
+    box-shadow: 0px 5px 4px 0px #cccccc;
 }
 
 /* Basic Artifact Card
 ========================================================================== */
 
 #artifact-card {
-    text-align: center;
     p {
         margin: 2px 0;
     }
@@ -204,8 +211,36 @@ body {
     #artifact-img {
         width: 100px;
         height: 100px;
-        border-radius: 10px;
+        border-radius: 4px;
         object-fit: cover;
+    }
+}
+
+/* Searchbar
+========================================================================== */
+
+#searchbar {
+    div {
+        display: flex;
+        input {
+            border-radius: 4px 0 0 4px;
+            border: transparent 1px solid;
+            padding: 8px 10px;
+        }
+
+        input:hover {
+            border: rgb(117, 159, 252) 1px solid;
+        }
+
+        button {
+            border-radius: 0 4px 4px 0;
+            border: none;
+            background-color: #eeeeee;
+        }
+
+        button:hover {
+            background-color: #d1d1d1;
+        }
     }
 }
